+++ src/main/java/com/example/samplebootapp/application/order/query/CartQueryService.java
@@ -1,22 +1,8 @@
 package com.example.samplebootapp.application.order.query;
 
-import com.example.samplebootapp.infrastructure.order.mapper.CartMapper;
-import java.util.Collections;
-import java.util.List;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 /** カート参照サービス. */
-@Service
-@Transactional(readOnly = true)
-public class CartQueryService {
-
-  private final CartMapper cartMapper;
-
-  @edu.umd.cs.findbugs.annotations.SuppressFBWarnings("EI_EXPOSE_REP2")
-  public CartQueryService(CartMapper cartMapper) {
-    this.cartMapper = cartMapper;
-  }
+/** カート参照サービス. */
+public interface CartQueryService {
 
   /**
    * カートを取得します.
@@ -24,15 +10,5 @@ public class CartQueryService {
    * @param userId ユーザーID
    * @return カートDTO
    */
-  public CartDto getCart(String userId) {
-    String cartId = cartMapper.findCartIdByUserId(userId);
-
-    if (cartId != null) {
-      List<CartItemDto> items = cartMapper.findCartItemDtos(cartId);
-      return new CartDto(cartId, items);
-    } else {
-      // カートが存在しない場合は空のレスポンスを返す
-      return new CartDto("", Collections.emptyList());
-    }
-  }
+  CartDto getCart(String userId);
 }
